# hjyserver/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/v1/device/askEd713RealData": {
            "post": {
                "description": "ask Ed713 device to send real data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ED713 device"
                ],
                "summary": "askEd713RealData",
                "parameters": [
                    {
                        "description": "ack Ed713 info",
                        "name": "in",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mdb.AskEd713RealDataReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "none"
                        }
                    }
                }
            }
        },
        "/v1/device/askX1RealData": {
            "post": {
                "description": "ask X1 device to send real data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "X1 device"
                ],
                "summary": "askX1RealData",
                "parameters": [
                    {
                        "description": "ack X1 info",
                        "name": "in",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mdb.AskX1RealDataReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "none"
                        }
                    }
                }
            }
        },
        "/v1/device/cleanX1Event": {
            "post": {
                "description": "clean event from X1 device",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "X1 device"
                ],
                "summary": "cleanX1Event",
                "parameters": [
                    {
                        "description": "clean X1 event",
                        "name": "in",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mdb.CleanX1EventReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "none"
                        }
                    }
                }
            }
        },
...
        "/v1/device/improveDisturbed": {
            "post": {
                "description": "import x1 device disturbed data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "X1 device"
                ],
                "summary": "improveDisturbed",
                "parameters": [
                    {
                        "description": "X1 distrube switch",
                        "name": "in",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mdb.ImproveDisturbedReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "none"
                        }
                    }
                }
            }
        },
...
        "/v1/device/queryX1RealDataJson": {
            "get": {
                "description": "query X1 real data in json format",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "X1 device"
                ],
                "summary": "queryX1RealDataJson",
                "parameters": [
                    {
                        "type": "string",
                        "description": "device mac address",
                        "name": "mac",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "create date",
                        "name": "create_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.X1RealDataOrigin"
                        }
                    }
                }
            }
        },
        "/v1/device/queryX1SleepReportJson": {
            "get": {
                "description": "query X1 sleep report in json format",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "X1 device"
                ],
                "summary": "queryX1SleepReportJson",
                "parameters": [
                    {
                        "type": "string",
                        "description": "device mac address",
                        "name": "mac",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "create date",
                        "name": "create_date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mysql.X1DayReportOrigin"
                        }
                    }
                }
            }
        },
...
        "/v1/device/sleepX1Switch": {
            "post": {
                "description": "change sleep switch of X1 device",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "X1 device"
                ],
                "summary": "sleepX1Switch",
                "parameters": [
                    {
                        "description": "X1 switch",
                        "name": "in",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mdb.SleepX1SwitchReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "none"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "mdb.AskEd713RealDataReq": {
            "type": "object",
            "properties": {
                "freq": {
                    "description": "发送频率\nrequired: false",
                    "type": "integer"
                },
                "keep_push": {
                    "description": "是否保持推送\nrequired: true\nenum: 0,1",
                    "type": "integer"
                },
                "mac": {
                    "description": "设备mac地址\nrequired: true",
                    "type": "string"
                }
            }
        },
        "mdb.AskX1RealDataReq": {
            "type": "object",
            "properties": {
                "freq": {
                    "description": "发送频率\nrequired: false",
                    "type": "integer"
                },
                "keep_push": {
                    "description": "是否保持推送\nrequired: true\nenum: 0,1",
                    "type": "integer"
                },
                "mac": {
                    "description": "设备mac地址\nrequired: true",
                    "type": "string"
                }
            }
        },
        "mdb.CleanX1EventReq": {
            "type": "object",
            "properties": {
                "mac": {
                    "description": "设备mac地址\nrequired: true",
                    "type": "string"
                }
            }
        },
...
        "mdb.ImproveDisturbedReq": {
            "type": "object",
            "properties": {
                "mac": {
                    "description": "设备mac地址\nrequired: true",
                    "type": "string"
                },
                "switch": {
                    "description": "开关状态\nrequired: true\nenum: 0,1",
                    "type": "integer"
                }
            }
        },
...
        "mdb.SleepX1SwitchReq": {
            "type": "object",
            "properties": {
                "mac": {
                    "description": "设备mac地址\nrequired: true",
                    "type": "string"
                },
                "switch": {
                    "description": "开关状态\nrequired: true\nenum: 0,1",
                    "type": "integer"
                }
            }
        },
...
        "mysql.X1DayReportOrigin": {
            "type": "object",
            "required": [
                "id",
                "mac"
            ],
            "properties": {
                "create_time": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "mac": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "mysql.X1RealDataOrigin": {
            "type": "object",
            "required": [
                "id",
                "mac"
            ],
            "properties": {
                "create_time": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "mac": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        }
    }
}